relocate FieldHint to common
